+++ tests/runScriptAnalyser.ps1
@@ -1,0 +1,1 @@
+Invoke-ScriptAnalyzer -Path "..\src\Dynamics365Configuration" -Recurse -ExcludeRule PSUseApprovedVerbs, PSUseDeclaredVarsMoreThanAssignments, PSAvoidUsingWMICmdlet